refactor(agent): extract worker domain from loop
