+++ packages/agent/src/loop/execution-supervisor.ts
@@ -44,8 +44,8 @@ import type {
   OrchestratorConfig,
   OrchestratorRunInput,
   ToolExecutor,
-} from "./run-worker";
-import { RunWorker } from "./run-worker";
+} from "../worker/run-worker";
+import { RunWorker } from "../worker/run-worker";
 
 const DEFAULT_DISPATCH_TIMEOUT_MS = 5 * 60 * 1000;
 const DEFAULT_MAX_SUBAGENT_DEPTH = 3;
