+++ packages/agent/src/loop/execution-types.ts
@@ -1,5 +1,5 @@
 import type { AgentDefinition } from "../agent";
-import type { OrchestratorRunInput, SessionMode } from "./run-worker";
+import type { OrchestratorRunInput, SessionMode } from "../worker/run-worker";
 
 export type SupervisorDecision = "local" | "spawn" | "join" | "finish";
