+++ packages/agent/src/loop/index.ts
@@ -1,51 +1,6 @@
-// Loop module - Phase 2 implementation
-// Event routing, dispatching, and concurrency control
-
-export { type ConcurrencyConfig, ConcurrencyGate } from "./concurrency";
-
-export {
-  type PermissionLevel,
-  type PermissionDecision,
-  type PermissionContext,
-  PermissionGate,
-} from "./permission";
-
-export {
-  type RunBudget,
-  type RunState,
-  type BudgetStatus,
-  RunSupervisor,
-} from "./run-supervisor";
-
-export type {
-  OrchestratorConfig,
-  OrchestrationResult,
-  OrchestrationState,
-  ToolExecutor,
-  SessionMode,
-  OrchestratorRunInput,
-} from "./run-worker";
-
-export { RunWorker } from "./run-worker";
+// Loop module — conversation/execution supervisors, DAG execution, file locking
 
 export { FileLock } from "./file-lock";
-
-export { type DLQEntry, DeadLetterQueue } from "./dlq";
-
-export {
-  type SummaryTemplate,
-  type SummaryData,
-  SummaryDelivery,
-} from "./summary";
-
-export {
-  type EventMetadata,
-  type QueueMetrics,
-  type RunMetrics,
-  Observability,
-} from "./observability";
-
-export { type AuditEntry, AuditLog } from "./audit";
 
 export {
   ConversationSupervisor,
@@ -96,11 +51,3 @@ export {
   requestHandoffDocument,
   rotateAgent,
 } from "./execution-review";
-
-export {
-  resolveAgentDefinition,
-  resolveLLM,
-  resolveToolExecutor,
-  resolveAgentForWorker,
-  fallbackToolExecutor,
-} from "./agent-resolution";
