+++ packages/agent/src/worker/agent-resolution.ts
