+++ packages/agent/src/worker/permission.ts
